Passing from text files to custom class

+++ src/Accounts.java
@@ -1,15 +1,15 @@
 public class Accounts {
-    private double money;
+    private long money;
 
-    public Accounts(double money) {
+    public Accounts(long money) {
         this.money = money;
     }
 
-    public double getMoney() {
+    public long getMoney() {
         return money;
     }
 
-    public void setMoney(double money) {
+    public void setMoney(long money) {
         this.money = money;
     }
 }
